Add doc comments, use min(property:) for mostCompact
