+++ Package.resolved
@@ -24,8 +24,8 @@
         "repositoryURL": "https://github.com/dn-m/Structure",
         "state": {
           "branch": null,
-          "revision": "16221b5068b7c3e52066929cb0a943d3ccc3f149",
-          "version": "0.6.0"
+          "revision": "08397f84bd17a5dc5adbd6864508ae0c36e1531b",
+          "version": "0.7.0"
         }
       }
     ]
